Fix editing behavior after a merge.
